+++ visualization/app/codeCharta/ui/dialog/dialog.changelog.po.ts
@@ -1,0 +1,6 @@
+export class DialogChangelogPageObject {
+	async doSomething() {
+		await page.waitForSelector("CSS_SELECTOR")
+		return page.$eval("A > CSS_SELECTOR", element => element["someAttribute"])
+	}
+}